SI-6502 Moving methods concerned with the state of Global from IMain to Global
