+++ src/compiler/scala/tools/nsc/Global.scala
@@ -8,6 +8,7 @@ package tools
 package nsc
 
 import java.io.{ File, FileOutputStream, PrintWriter, IOException, FileNotFoundException }
+import java.net.URL
 import java.nio.charset.{ Charset, CharsetDecoder, IllegalCharsetNameException, UnsupportedCharsetException }
 import scala.compat.Platform.currentTime
 import scala.collection.{ mutable, immutable }
@@ -839,6 +840,40 @@ class Global(var currentSettings: Settings, var reporter: Reporter)
       if (res.nonEmpty && res.head._2 == value) res
       else ((ph, value)) :: res
     } reverse
+  }
+
+  // ------------ REPL utilities ---------------------------------
+
+  /** Extend classpath of `platform` and rescan updated packages. */
+  def extendCompilerClassPath(urls: URL*): Unit = {
+    val newClassPath = mergeUrlsIntoClassPath(platform, urls: _*)
+    platform.currentClassPath = Some(newClassPath)
+    // Reload all specified jars into this compiler instance
+    invalidateClassPathEntries(urls.map(_.getPath): _*)
+  }
+
+  /** Merge classpath of `platform` and `urls` into merged classpath */
+  def mergeUrlsIntoClassPath(platform: JavaPlatform, urls: URL*): MergedClassPath[AbstractFile] = {
+    // Collect our new jars/directories and add them to the existing set of classpaths
+    val prevEntries = platform.classPath match {
+      case mcp: MergedClassPath[AbstractFile] => mcp.entries
+      case cp:  ClassPath[AbstractFile]       => List(cp)
+    }
+    val allEntries = (prevEntries ++
+      urls.map(url => platform.classPath.context.newClassPath(
+          if (url.getProtocol == "file") {
+            val f = new File(url.getPath)
+            if (f.isDirectory) io.AbstractFile.getDirectory(f)
+            else io.AbstractFile.getFile(f)
+          } else {
+            io.AbstractFile.getURL(url)
+          }
+        )
+      )
+    ).distinct
+
+    // Combine all of our classpaths (old and new) into one merged classpath
+    new MergedClassPath(allEntries, platform.classPath.context)
   }
 
   // ------------ Invalidations ---------------------------------
